requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
google-generativeai==0.4.0
pypdf==3.17.1
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0
ruff==0.1.9
